Added lookup indices when creating a FlatTree.
In the case were the produced FlatTree is smaller that the original tree,
it might be usefull to have an associative table between each new node and
the original node.

`new_with_lookup_table` contructor gives you back a buffer of the exact same
length as the FlatTree produced buffer containing indices.
Each of those index refer to the equivalent original tree flatten by FlatTree
with no node excluded.

This is currently useful, as LayoutBuffer is a flatten version of the StyledNode
tree, and both RenderBuffer and FocusBuffer depend on LayoutBuffer.